% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie-index.R
\name{EI_index}
\alias{EI_index}
\alias{EI_index.list}
\alias{EI_index.network}
\alias{EI_index.igraph}
\alias{EI_index.matrix}
\alias{EI_index.default}
\title{Krackhardt and Sterns' EI index}
\usage{
EI_index(net, ..., depth = 0)

\method{EI_index}{list}(net, vname, ..., depth = 0)

\method{EI_index}{network}(net, vname, ...)

\method{EI_index}{igraph}(net, vname, ...)

\method{EI_index}{matrix}(net, v, ...)

\method{EI_index}{default}(x, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The list objects can even be mixed, meaning that some objects are of type network, some of type
igraph, etc. The networks can either be directed or undirected. In case of a directed network it
is recommended to specify whether the indices should be calculated based on the incoming network
ties or outgoing network ties.}

\item{...}{Potential additional paramters. .}

\item{depth}{this paramter should not be manipulated! It controls the recursion depth of the function
in case net is of type list.}

\item{vname}{character variable indicating which variable should be used for grouping. This parameter
can only be used if net is of type igraph, network or list. Please note, that in case you are
using the list input, the variable has to be present in each of the list objects. This parameter
can only be used if net is of type igraph, network or list.}

\item{v}{vector of node attributes. This parameter is only relevant if net is of type matrix. Otherwise
this parameter should be ignored and left as it is.}
}
\value{
a numeric vector of a normalized difderence for each individual between
the number of in and outgroup ties.
}
\description{
`EI_index()` calculates an inverse homophily index based on a given network and node attribute.
}
\details{
During the developement of the IOIndex this index was used for validation. The index was developed
 by Krackhardt & Stern (1988) DOI:  DOI: 10.2307/2786835. Thus both indices, the IOIndex and the
 EI-index, should usually be moderatedly and negatively correlated. However, as the EI-Index
 is known to be sensitive to group size, the association stregth with the IOIndex
 should be expected to be smaller than for the other indices in this package.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type network

\item \code{default}: error case
}}

\examples{
# Please consult the documentation for the social_closing, social_opneing, etc. indices

}
