% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{IE_index}
\alias{IE_index}
\alias{IE_index.list}
\alias{IE_index.network}
\alias{IE_index.igraph}
\alias{IE_index.matrix}
\alias{IE_index.default}
\title{TITLE}
\usage{
IE_index(net, ..., depth = 0)

\method{IE_index}{list}(net, vname, ..., depth = 0)

\method{IE_index}{network}(net, vname, ...)

\method{IE_index}{igraph}(net, vname, ...)

\method{IE_index}{matrix}(net, v, ...)

\method{IE_index}{default}(x, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The list objects can even be mixed, meaning that some objects are of type network, some of type
igraph, etc. The networks can either be directed or undirected. In case of a directed network it
is recommended to specify whether the indices should be calculated based on the incoming network
ties or outgoing network ties.}

\item{...}{Potential additional paramters. .}

\item{depth}{this paramter should not be manipulated! It controls the recursion depth of the function
in case net is of type list.}

\item{vname}{character variable indicating which variable should be used for grouping. This parameter
can only be used if net is of type igraph, network or list. Please note, that in case you are
using the list input, the variable has to be present in each of the list objects. This parameter
can only be used if net is of type igraph, network or list.}

\item{v}{vector of node attributes. This parameter is only relevant if net is of type matrix. Otherwise
this parameter should be ignored and left as it is.}
}
\value{
a numeric vector of ???
}
\description{
`IE_index()` calculates a homophily index based on a given network and node attribute.
This index was invented by McCormick, Cappella, Hughes, Gallagher (2015). DOI: 10.1177/0272431614547051. During the developement of the IOIndex this index was used for validation.
Thus both indices, the IOIndex and McCormick's index, should usually be moderatedly correlated.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type network

\item \code{default}: error case
}}

