% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-indices.R
\name{naive_closing}
\alias{naive_closing}
\title{Naive in- and outgroup index}
\usage{
naive_closing(net, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The list objects can even be mixed, meaning that some objects are of type network, some of type
igraph, etc. The networks can either be directed or undirected. In case of a directed network it
is recommended to specify whether the indices should be calculated based on the incoming network
ties or outgoing network ties.}

\item{...}{Potential additional paramters. .}
}
\description{
This index was described and used in various papers (see e.g. Reitz, Asendorpf & Motti-Stefanidi, 2015, DOI: 10.1177/0165025414567008) For this index a ratio
between the actual number relationships to members of the in- or outgroup is compared to
the maximum possible number of connections within the in- or outgroup
(aka the size of the in- or outgroup).
}
