% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{naive_opening}
\alias{naive_opening}
\alias{naive_opening.list}
\alias{naive_opening.network}
\alias{naive_opening.igraph}
\alias{naive_opening.matrix}
\alias{naive_opening.default}
\alias{naive_closing.list}
\alias{naive_closing.network}
\alias{naive_closing.igraph}
\alias{naive_closing.matrix}
\alias{naive_closing.default}
\title{Naive in- and outgroup index}
\usage{
naive_opening(net, ...)

\method{naive_opening}{list}(net, vname, depth = 0)

\method{naive_opening}{network}(net, vname)

\method{naive_opening}{igraph}(net, vname)

\method{naive_opening}{matrix}(net, v)

\method{naive_opening}{default}(x, depth = 0, ...)

\method{naive_closing}{list}(net, vname, depth = 0)

\method{naive_closing}{network}(net, vname)

\method{naive_closing}{igraph}(net, vname)

\method{naive_closing}{matrix}(net, v)

\method{naive_closing}{default}(net, ...)
}
\description{
This index was described and used in various papers (see e.g. Reitz, Asendorpf & Motti-Stefanidi, 2015, DOI: 10.1177/0165025414567008) For this index a ratio
between the actual number relationships to members of the in- or outgroup is compared to
the maximum possible number of connections within the in- or outgroup
(aka the size of the in- or outgroup).
}
\details{
#' @inheritParams social_closing
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix

\item \code{default}: error case

\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix

\item \code{default}: error case
}}

