% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{social_all}
\alias{social_all}
\title{Social closing and opening coefficients}
\usage{
social_all(net, vname)
}
\arguments{
\item{net}{Network of type igraph, network or a list of objects of these datatypes. The
type matrix is not allowed in this function.}

\item{vname}{character variable indicating the . This parameter can only be used if net is of
type igraph, network or list.}
}
\value{
a three column data.frame with the social closing, opening and integration coefficients
per person in the network.
}
\description{
`social_all()` calculates the social closing, opening and
integration coefficients for a given network and node attribute.
}
