% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{social_closing}
\alias{social_closing}
\alias{social_closing.list}
\alias{social_closing.network}
\alias{social_closing.igraph}
\alias{social_closing.matrix}
\alias{social_closing.default}
\title{Social closing coefficient}
\usage{
social_closing(net, ..., depth = 0)

\method{social_closing}{list}(net, vname, depth = 0)

\method{social_closing}{network}(net, vname, ...)

\method{social_closing}{igraph}(net, vname, ...)

\method{social_closing}{matrix}(net, v, ...)

\method{social_closing}{default}(x, depth = 0, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The network can either be directed or undirected.}

\item{...}{Addtional paramters}

\item{depth}{this paramter should not be manipulated! It controls the recursion depth of the function
In case net is of type list.}

\item{vname}{character variable indicating the . This parameter can only be used if net is of
type igraph, network or list.}

\item{v}{vector of node attributes.}
}
\value{
a numeric vector of social closinig indices for each indvidual in the network.
}
\description{
`social_closing()` calculates the social closing coefficient for a given network and node attribute.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix

\item \code{default}: error case
}}

\examples{

data("faux.magnolia.high", package = "ergm")
fmh <- faux.magnolia.high ; rm(faux.magnolia.high)
social_closing(fmh, "Grade")

x <- network::as.sociomatrix(fmh)
v <- network::get.vertex.attribute(fmh, "Grade")
social_closing(x, v)

gg <- igraph::graph_from_adjacency_matrix(x)
igraph::vertex_attr(gg, "Grade") <- v

social_closing(list(fmh, fmh), "Grade")
social_closing(list(gg, fmh), "Grade")
}
