% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{social_integration}
\alias{social_integration}
\alias{social_integration.network}
\alias{social_integration.igraph}
\alias{social_integration.matrix}
\title{Social integration coefficient}
\usage{
social_integration(net, ...)

\method{social_integration}{network}(net, vname)

\method{social_integration}{igraph}(net, vname)

\method{social_integration}{matrix}(net, v)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The network can either be directed or undirected.}

\item{...}{Addtional paramters}

\item{vname}{character variable indicating the . This parameter can only be used if net is of
type igraph, network or list.}

\item{v}{vector of node attributes.}
}
\description{
`social_integration()` calculates the social integration coefficient for a given network and node attribute.
}
\section{Methods (by class)}{
\itemize{
\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix
}}

