% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{naive_closing}
\alias{naive_closing}
\title{Naive in- and outgroup index}
\usage{
naive_closing(net, ...)
}
\description{
This index was described and used in various papers (see e.g. Reitz, Asendorpf & Motti-Stefanidi, 2015, DOI: 10.1177/0165025414567008) For this index a ratio
between the actual number relationships to members of the in- or outgroup is compared to
the maximum possible number of connections within the in- or outgroup
(aka the size of the in- or outgroup).
}
