% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-indices.R
\name{naive_opening}
\alias{naive_opening}
\alias{naive_opening.list}
\alias{naive_opening.network}
\alias{naive_opening.igraph}
\alias{naive_opening.matrix}
\alias{naive_opening.default}
\alias{naive_closing.list}
\alias{naive_closing.network}
\alias{naive_closing.igraph}
\alias{naive_closing.matrix}
\alias{naive_closing.default}
\title{Naive in- and outgroup index}
\usage{
naive_opening(net, ...)

\method{naive_opening}{list}(net, vname, depth = 0)

\method{naive_opening}{network}(net, vname)

\method{naive_opening}{igraph}(net, vname)

\method{naive_opening}{matrix}(net, v)

\method{naive_opening}{default}(x, depth = 0, ...)

\method{naive_closing}{list}(net, vname, depth = 0)

\method{naive_closing}{network}(net, vname)

\method{naive_closing}{igraph}(net, vname)

\method{naive_closing}{matrix}(net, v)

\method{naive_closing}{default}(net, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The list objects can even be mixed, meaning that some objects are of type network, some of type
igraph, etc. The networks can either be directed or undirected. In case of a directed network it
is recommended to specify whether the indices should be calculated based on the incoming network
ties or outgoing network ties.}

\item{...}{Potential additional paramters. .}

\item{vname}{character variable indicating which variable should be used for grouping. This parameter
can only be used if net is of type igraph, network or list. Please note, that in case you are
using the list input, the variable has to be present in each of the list objects. This parameter
can only be used if net is of type igraph, network or list.}

\item{depth}{this paramter should not be manipulated! It controls the recursion depth of the function
in case net is of type list.}

\item{v}{vector of node attributes. This parameter is only relevant if net is of type matrix. Otherwise
this parameter should be ignored and left as it is.}
}
\value{
a numeric vector of real valued ratios for each indvidual in the network.
}
\description{
This index was described and used in various papers (see e.g. Reitz, Asendorpf & Motti-Stefanidi, 2015, DOI: 10.1177/0165025414567008) For this index a ratio
between the actual number relationships to members of the in- or outgroup is compared to
the maximum possible number of connections within the in- or outgroup
(aka the size of the in- or outgroup).
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix

\item \code{default}: error case

\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix

\item \code{default}: error case
}}

\examples{
# Please consult the documentation for the social_closing, social_opneing, etc. indices

}
