% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{social_all}
\alias{social_all}
\title{Social closing and opening coefficients}
\usage{
social_all(net, vname)
}
\arguments{
\item{net}{Network of either type igraph, network or a list of objects of any of these datatypes.
The list objects can even be mixed, meaing that some objects are of type network, some of type
igraph, etc. However, the type matrix is not allowed in this function.}

\item{vname}{character variable indicating the variable that is used for grouping. This parameter
can only be used if net is of type igraph, network or list. Please note, that in case you are
using the list input, the variable has to be present in each of the list objects.}
}
\value{
a data.frame with three columns for the social closing, opening and integration coefficients
per person in the given network(s).
}
\description{
`social_all()` calculates the social closing, opening and
integration coefficients for a given network and set of node attributes.
}
