% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{social_opening}
\alias{social_opening}
\alias{social_opening.network}
\alias{social_opening.igraph}
\alias{social_opening.matrix}
\title{Social opening coefficient}
\usage{
social_opening(net, ...)

\method{social_opening}{network}(net, vname)

\method{social_opening}{igraph}(net, vname)

\method{social_opening}{matrix}(net, v)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The network can either be directed or undirected.}

\item{...}{Addtional paramters}

\item{vname}{character variable indicating the . This parameter can only be used if net is of
type igraph, network or list.}

\item{v}{vector of node attributes.}
}
\description{
`social_opening()` calculates the social opening coefficient for a given network and node attribute.
}
\section{Methods (by class)}{
\itemize{
\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type matrix
}}

