% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yules-q.R
\name{Yules_q}
\alias{Yules_q}
\alias{Yules_q.list}
\alias{Yules_q.network}
\alias{Yules_q.igraph}
\alias{Yules_q.matrix}
\alias{Yules_q.default}
\title{Yule's Q}
\usage{
Yules_q(net, ..., depth = 0)

\method{Yules_q}{list}(net, vname, ..., depth = 0)

\method{Yules_q}{network}(net, vname, ...)

\method{Yules_q}{igraph}(net, vname, ...)

\method{Yules_q}{matrix}(net, v, ...)

\method{Yules_q}{default}(net, ...)
}
\arguments{
\item{net}{Network of type matrix, igraph, network or a list of objects of these datatypes.
The list objects can even be mixed, meaning that some objects are of type network, some of type
igraph, etc. The networks can either be directed or undirected. In case of a directed network it
is recommended to specify whether the indices should be calculated based on the incoming network
ties or outgoing network ties.}

\item{...}{Potential additional paramters. .}

\item{depth}{this paramter should not be manipulated! It controls the recursion depth of the function
in case net is of type list.}

\item{vname}{character variable indicating which variable should be used for grouping. This parameter
can only be used if net is of type igraph, network or list. Please note, that in case you are
using the list input, the variable has to be present in each of the list objects. This parameter
can only be used if net is of type igraph, network or list.}

\item{v}{vector of node attributes. This parameter is only relevant if net is of type matrix. Otherwise
this parameter should be ignored and left as it is.}
}
\value{
a numeric vector of a normalized difference for each individual between
the number of in and outgroup ties and non-ties.
}
\description{
`Yules_q()` calculates a homophily index based on a given network and node attribute.
}
\details{
Yules_q can be used as an alternative for the well known EI index by Krackhardt and Stern, which
is also implemeted in this package. In contrast to the EI Index, Yule's Q also accounts for
the distribution of nonn -ties. For further information on this index and its comparison to the EI
index see Borgatti, S. P., Everett, M. G., & Johnson, Jeffrey C. (2013). Analyzing social networks.
Los Angeles, et al.: SAGE Publications. The EI index by Krackhardt & Stern (1988)
DOI: 10.2307/2786835. was used during the developement of the IOIindex as a validation measure. Yule's Q
was not during the developement of the package but in principle it could also have served as a
validation for our new index.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: net is of type list

\item \code{network}: net is of type network

\item \code{igraph}: net is of type igraph

\item \code{matrix}: net is of type network

\item \code{default}: error case
}}

\examples{
# Please consult the documentation for the social_closing, social_opneing, etc. indices

}
